Support installing latest with EnsurePackage
Originally EnsurePackage was written for the older go get behavior. Now
that we are using go install, the required syntax is a bit different,
for example when a version is not specified, we should use @latest. I've
also made it easier for people to specify the version both with and
without the v prefix, and allow for them specifying either an empty
version or latest.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>