Use version range when checking if installed
When we check if a command of a specific version is available, allow the
user to specify a version range, e.g. 2.x or ^1.2.4

When installing a package, default the version range check to using the
default version as the minimum, e.g. ^defaultVersion. This allows any
thing from the default version to the next major version.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>