Add EnsurePackageWith/InstallPackageWith
EnsurePackageWith replaces EnsurePackage, which is now marked as
deprecated. Same for InstallPackage.

The With variation accepts an options struct so that we can
add functionality without modifying the function signature.
You can now specify a destination directory instead of installing into
GOPATH/bin. This is useful for installing build/dev tools into ./bin and
not messing with the system installation on dev machines.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>